Make sure dnssd constants and browsedelegate is tracked by gn
