+++ src/app/server/BUILD.gn
@@ -55,6 +55,7 @@ static_library("server") {
     "${chip_root}/src/app/icd/server:observer",
     "${chip_root}/src/lib/address_resolve",
     "${chip_root}/src/lib/dnssd",
+    "${chip_root}/src/lib/dnssd:naming",
     "${chip_root}/src/messaging",
     "${chip_root}/src/platform",
     "${chip_root}/src/protocols",
